# openapi_contract.yaml
openapi: 3.0.3
info:
  title: To-Do Backend API
  version: 1.0.0
  description: |
    Formal OpenAPI-style contract for the FastAPI to-do backend.
    This document reflects the currently implemented routes and payload shapes.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [users]
      summary: List users (admin only)
      operationId: getUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Error403'
  /users/{user_id}:
    put:
      tags: [users]
      summary: Update user (admin only)
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/Error404'
    delete:
      tags: [users]
      summary: Delete user (admin only)
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
  /users/remediate-passwords:
    post:
      tags: [users]
      summary: Remediate invalid password hashes (admin only)
      operationId: remediatePasswords
      parameters:
        - in: query
          name: dry_run
          required: false
          schema:
            type: boolean
            default: true
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: Remediation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordRemediationResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserIdPath:
      in: path
      name: user_id
      required: true
      schema:
        type: integer
...
  responses:
...
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [message, code, dev_message, request_id, path]
      properties:
        message:
          type: string
        code:
          type: string
        dev_message:
          type: string
        request_id:
          type: string
        path:
          type: string
        details:
          type: array
          items: {}
...
    DepartmentReference:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    UserResponse:
      type: object
      required: [id, name, email, role, departments]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        departments:
          type: array
          items:
            $ref: '#/components/schemas/DepartmentReference'
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        role:
          type: string
    PasswordRemediationUser:
      type: object
      required: [user_id, username, temporary_password]
      properties:
        user_id:
          type: integer
        username:
          type: string
        email:
          type: string
          nullable: true
        temporary_password:
          type: string
    PasswordRemediationResponse:
      type: object
      required: [processed_users, affected_users]
      properties:
        processed_users:
          type: integer
        affected_users:
          type: array
          items:
            $ref: '#/components/schemas/PasswordRemediationUser'
...